requirements.txt
django
django-import-export
psycopg2-binary
python-dotenv
openpyxl